// src/app/model/persona.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Autorizacion } from './autorizacion';

export interface Persona { 
    autorizacion?: Autorizacion;
    cedula?: string;
    idPersona?: number;
    nombre?: string;
    parentesco?: string;
    telefono?: string;
    tipo?: string;
}